@@ -1,8 +1,8 @@
 // Export base classes and interfaces
-export * from './base-embedding';
+export * from './base-embedding'
 
+export * from './gemini-embedding'
+export * from './ollama-embedding'
 // Implementation class exports
-export * from './openai-embedding';
-export * from './voyageai-embedding';
-export * from './ollama-embedding';
-export * from './gemini-embedding'; 
+export * from './openai-embedding'
+export * from './voyageai-embedding'
