@@ -1,14 +1,14 @@
 export interface SearchQuery {
-    term: string;
-    includeContent?: boolean;
-    limit?: number;
+  term: string
+  includeContent?: boolean
+  limit?: number
 }
 
 export interface SemanticSearchResult {
-    content: string;
-    relativePath: string;
-    startLine: number;
-    endLine: number;
-    language: string;
-    score: number;
+  content: string
+  relativePath: string
+  startLine: number
+  endLine: number
+  language: string
+  score: number
 }
