@@ -1,1 +1,1 @@
-export { EnvManager, envManager } from './env-manager'; 
+export { EnvManager, envManager } from './env-manager'
