@@ -1,5 +1,5 @@
 export const config = {
-    apiUrl: 'https://api.example.com',
-    timeout: 5000,
-    retries: 3
-};
+  apiUrl: 'https://api.example.com',
+  timeout: 5000,
+  retries: 3,
+}
