@@ -1,9 +1,9 @@
 export class Logger {
-    log(message: string): void {
-        console.log(`[LOG] ${message}`);
-    }
+  log(message: string): void {
+    console.log(`[LOG] ${message}`)
+  }
 
-    error(message: string): void {
-        console.error(`[ERROR] ${message}`);
-    }
+  error(message: string): void {
+    console.error(`[ERROR] ${message}`)
+  }
 }
