@@ -1,289 +1,289 @@
-import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import * as path from 'path';
-import { TestContextBuilder } from '../doubles/test-context-builder';
-import { FakeVectorDatabase } from '../doubles/fake-vector-database';
-import { FakeEmbedding } from '../doubles/fake-embedding';
-import { Context } from '../../src/context';
+import type { Context } from '../../src/context'
+import * as path from 'node:path'
+import { afterEach, beforeEach, describe, expect, it } from 'vitest'
+import { FakeEmbedding } from '../doubles/fake-embedding'
+import { FakeVectorDatabase } from '../doubles/fake-vector-database'
+import { TestContextBuilder } from '../doubles/test-context-builder'
 
-describe('Indexing Workflow Integration', () => {
-    let context: Context;
-    let fakeDb: FakeVectorDatabase;
-    let fakeEmbedding: FakeEmbedding;
-    let fixturesPath: string;
+describe('indexing Workflow Integration', () => {
+  let context: Context
+  let fakeDb: FakeVectorDatabase
+  let fakeEmbedding: FakeEmbedding
+  let fixturesPath: string
 
-    beforeEach(() => {
-        // Create test doubles
-        fakeDb = new FakeVectorDatabase({ address: 'test' });
-        fakeEmbedding = new FakeEmbedding(128);
+  beforeEach(() => {
+    // Create test doubles
+    fakeDb = new FakeVectorDatabase({ address: 'test' })
+    fakeEmbedding = new FakeEmbedding(128)
 
-        // Create context with test doubles
-        context = new TestContextBuilder()
-            .withEmbedding(fakeEmbedding)
-            .withVectorDatabase(fakeDb)
-            .build();
+    // Create context with test doubles
+    context = new TestContextBuilder()
+      .withEmbedding(fakeEmbedding)
+      .withVectorDatabase(fakeDb)
+      .build()
 
-        // Path to test fixtures
-        fixturesPath = path.join(__dirname, '../fixtures/sample-codebase');
-    });
+    // Path to test fixtures
+    fixturesPath = path.join(__dirname, '../fixtures/sample-codebase')
+  })
 
-    afterEach(() => {
-        // Clean up test doubles
-        fakeDb.reset();
-        fakeEmbedding.reset();
-    });
+  afterEach(() => {
+    // Clean up test doubles
+    fakeDb.reset()
+    fakeEmbedding.reset()
+  })
 
-    describe('Basic Indexing', () => {
-        it('should index a codebase and create a collection', async () => {
-            // Act
-            const result = await context.indexCodebase(fixturesPath);
+  describe('basic Indexing', () => {
+    it('should index a codebase and create a collection', async () => {
+      // Act
+      const result = await context.indexCodebase(fixturesPath)
 
-            // Assert
-            expect(result.indexedFiles).toBeGreaterThan(0);
-            expect(result.totalChunks).toBeGreaterThan(0);
-            expect(result.status).toBe('completed');
+      // Assert
+      expect(result.indexedFiles).toBeGreaterThan(0)
+      expect(result.totalChunks).toBeGreaterThan(0)
+      expect(result.status).toBe('completed')
 
-            // Verify collection was created
-            const collectionName = context.getCollectionName(fixturesPath);
-            expect(await fakeDb.hasCollection(collectionName)).toBe(true);
-        });
+      // Verify collection was created
+      const collectionName = context.getCollectionName(fixturesPath)
+      expect(await fakeDb.hasCollection(collectionName)).toBe(true)
+    })
 
-        it('should generate embeddings for all chunks', async () => {
-            // Act
-            await context.indexCodebase(fixturesPath);
+    it('should generate embeddings for all chunks', async () => {
+      // Act
+      await context.indexCodebase(fixturesPath)
 
-            // Assert: Embedding provider was called
-            expect(fakeEmbedding.getCallCount()).toBeGreaterThan(0);
-            expect(fakeEmbedding.getEmbeddedTexts().length).toBeGreaterThan(0);
-        });
+      // Assert: Embedding provider was called
+      expect(fakeEmbedding.getCallCount()).toBeGreaterThan(0)
+      expect(fakeEmbedding.getEmbeddedTexts().length).toBeGreaterThan(0)
+    })
 
-        it('should store vectors in the database', async () => {
-            // Act
-            await context.indexCodebase(fixturesPath);
+    it('should store vectors in the database', async () => {
+      // Act
+      await context.indexCodebase(fixturesPath)
 
-            // Assert: Documents were inserted
-            const collectionName = context.getCollectionName(fixturesPath);
-            const documentCount = fakeDb.getCollectionDocumentCount(collectionName);
+      // Assert: Documents were inserted
+      const collectionName = context.getCollectionName(fixturesPath)
+      const documentCount = fakeDb.getCollectionDocumentCount(collectionName)
 
-            expect(documentCount).toBeGreaterThan(0);
+      expect(documentCount).toBeGreaterThan(0)
 
-            // Verify document structure
-            const documents = fakeDb.getStoredDocuments(collectionName);
-            expect(documents.length).toBeGreaterThan(0);
+      // Verify document structure
+      const documents = fakeDb.getStoredDocuments(collectionName)
+      expect(documents.length).toBeGreaterThan(0)
 
-            const firstDoc = documents[0];
-            expect(firstDoc).toHaveProperty('id');
-            expect(firstDoc).toHaveProperty('vector');
-            expect(firstDoc).toHaveProperty('content');
-            expect(firstDoc).toHaveProperty('relativePath');
-            expect(firstDoc).toHaveProperty('startLine');
-            expect(firstDoc).toHaveProperty('endLine');
-            expect(firstDoc).toHaveProperty('fileExtension');
+      const firstDoc = documents[0]
+      expect(firstDoc).toHaveProperty('id')
+      expect(firstDoc).toHaveProperty('vector')
+      expect(firstDoc).toHaveProperty('content')
+      expect(firstDoc).toHaveProperty('relativePath')
+      expect(firstDoc).toHaveProperty('startLine')
+      expect(firstDoc).toHaveProperty('endLine')
+      expect(firstDoc).toHaveProperty('fileExtension')
 
-            // Verify vector dimension matches embedding dimension
-            expect(firstDoc.vector.length).toBe(128);
-        });
+      // Verify vector dimension matches embedding dimension
+      expect(firstDoc.vector.length).toBe(128)
+    })
 
-        it('should index multiple files with different extensions', async () => {
-            // Act
-            const result = await context.indexCodebase(fixturesPath);
+    it('should index multiple files with different extensions', async () => {
+      // Act
+      const result = await context.indexCodebase(fixturesPath)
 
-            // Assert: Multiple files were indexed
-            expect(result.indexedFiles).toBeGreaterThanOrEqual(2);
+      // Assert: Multiple files were indexed
+      expect(result.indexedFiles).toBeGreaterThanOrEqual(2)
 
-            // Verify documents from different file types exist
-            const collectionName = context.getCollectionName(fixturesPath);
-            const documents = fakeDb.getStoredDocuments(collectionName);
+      // Verify documents from different file types exist
+      const collectionName = context.getCollectionName(fixturesPath)
+      const documents = fakeDb.getStoredDocuments(collectionName)
 
-            const fileExtensions = new Set(documents.map(doc => doc.fileExtension));
+      const fileExtensions = new Set(documents.map((doc) => doc.fileExtension))
 
-            expect(fileExtensions.has('.ts')).toBe(true);
-            expect(fileExtensions.has('.py')).toBe(true);
-        });
+      expect(fileExtensions.has('.ts')).toBe(true)
+      expect(fileExtensions.has('.py')).toBe(true)
+    })
 
-        it('should handle empty codebase gracefully', async () => {
-            // Arrange: Create context with empty directory
-            const emptyPath = path.join(__dirname, '../fixtures/empty-dir');
+    it('should handle empty codebase gracefully', async () => {
+      // Arrange: Create context with empty directory
+      const emptyPath = path.join(__dirname, '../fixtures/empty-dir')
 
-            // Act
-            const result = await context.indexCodebase(emptyPath);
+      // Act
+      const result = await context.indexCodebase(emptyPath)
 
-            // Assert
-            expect(result.indexedFiles).toBe(0);
-            expect(result.totalChunks).toBe(0);
-            expect(result.status).toBe('completed');
-        });
-    });
+      // Assert
+      expect(result.indexedFiles).toBe(0)
+      expect(result.totalChunks).toBe(0)
+      expect(result.status).toBe('completed')
+    })
+  })
 
-    describe('Progress Tracking', () => {
-        it('should call progress callback during indexing', async () => {
-            // Arrange
-            const progressUpdates: Array<{ phase: string; percentage: number }> = [];
-            const progressCallback = (progress: any) => {
-                progressUpdates.push({
-                    phase: progress.phase,
-                    percentage: progress.percentage,
-                });
-            };
+  describe('progress Tracking', () => {
+    it('should call progress callback during indexing', async () => {
+      // Arrange
+      const progressUpdates: Array<{ phase: string, percentage: number }> = []
+      const progressCallback = (progress: any) => {
+        progressUpdates.push({
+          phase: progress.phase,
+          percentage: progress.percentage,
+        })
+      }
 
-            // Act
-            await context.indexCodebase(fixturesPath, progressCallback);
+      // Act
+      await context.indexCodebase(fixturesPath, progressCallback)
 
-            // Assert
-            expect(progressUpdates.length).toBeGreaterThan(0);
+      // Assert
+      expect(progressUpdates.length).toBeGreaterThan(0)
 
-            // Verify progress phases
-            const phases = progressUpdates.map(u => u.phase);
-            expect(phases.some(p => p.includes('Preparing'))).toBe(true);
-            expect(phases.some(p => p.includes('Scanning'))).toBe(true);
+      // Verify progress phases
+      const phases = progressUpdates.map((u) => u.phase)
+      expect(phases.some((p) => p.includes('Preparing'))).toBe(true)
+      expect(phases.some((p) => p.includes('Scanning'))).toBe(true)
 
-            // Verify progress percentage increases
-            const firstPercentage = progressUpdates[0].percentage;
-            const lastPercentage = progressUpdates[progressUpdates.length - 1].percentage;
+      // Verify progress percentage increases
+      const firstPercentage = progressUpdates[0].percentage
+      const lastPercentage = progressUpdates[progressUpdates.length - 1].percentage
 
-            expect(lastPercentage).toBeGreaterThanOrEqual(firstPercentage);
-        });
+      expect(lastPercentage).toBeGreaterThanOrEqual(firstPercentage)
+    })
 
-        it('should report 100% progress on completion', async () => {
-            // Arrange
-            let lastProgress: any = null;
-            const progressCallback = (progress: any) => {
-                lastProgress = progress;
-            };
+    it('should report 100% progress on completion', async () => {
+      // Arrange
+      let lastProgress: any = null
+      const progressCallback = (progress: any) => {
+        lastProgress = progress
+      }
 
-            // Act
-            await context.indexCodebase(fixturesPath, progressCallback);
+      // Act
+      await context.indexCodebase(fixturesPath, progressCallback)
 
-            // Assert
-            expect(lastProgress).not.toBeNull();
-            expect(lastProgress.percentage).toBe(100);
-        });
-    });
+      // Assert
+      expect(lastProgress).not.toBeNull()
+      expect(lastProgress.percentage).toBe(100)
+    })
+  })
 
-    describe('File Extension Filtering', () => {
-        it('should respect supported extensions configuration', async () => {
-            // Arrange: Create context that only supports .ts files
-            const customContext = new TestContextBuilder()
-                .withEmbedding(fakeEmbedding)
-                .withVectorDatabase(fakeDb)
-                .withSupportedExtensions(['.ts'])
-                .build();
+  describe('file Extension Filtering', () => {
+    it('should respect supported extensions configuration', async () => {
+      // Arrange: Create context that only supports .ts files
+      const customContext = new TestContextBuilder()
+        .withEmbedding(fakeEmbedding)
+        .withVectorDatabase(fakeDb)
+        .withSupportedExtensions(['.ts'])
+        .build()
 
-            // Act
-            await customContext.indexCodebase(fixturesPath);
+      // Act
+      await customContext.indexCodebase(fixturesPath)
 
-            // Assert: Only .ts files were indexed
-            const collectionName = customContext.getCollectionName(fixturesPath);
-            const documents = fakeDb.getStoredDocuments(collectionName);
+      // Assert: Only .ts files were indexed
+      const collectionName = customContext.getCollectionName(fixturesPath)
+      const documents = fakeDb.getStoredDocuments(collectionName)
 
-            const fileExtensions = new Set(documents.map(doc => doc.fileExtension));
+      const fileExtensions = new Set(documents.map((doc) => doc.fileExtension))
 
-            expect(fileExtensions.has('.ts')).toBe(true);
+      expect(fileExtensions.has('.ts')).toBe(true)
 
-            // If Python files were indexed, this is a bug
-            if (fileExtensions.has('.py')) {
-                console.log('WARNING: Python files were indexed despite filter. This may be expected behavior.');
-            }
-        });
-    });
+      // If Python files were indexed, this is a bug
+      if (fileExtensions.has('.py')) {
+        console.log('WARNING: Python files were indexed despite filter. This may be expected behavior.')
+      }
+    })
+  })
 
-    describe('Hybrid vs Regular Collection', () => {
-        it('should create hybrid collection by default', async () => {
-            // Act
-            await context.indexCodebase(fixturesPath);
+  describe('hybrid vs Regular Collection', () => {
+    it('should create hybrid collection by default', async () => {
+      // Act
+      await context.indexCodebase(fixturesPath)
 
-            // Assert
-            const collectionName = context.getCollectionName(fixturesPath);
-            expect(fakeDb.isHybridCollection(collectionName)).toBe(true);
-        });
-    });
+      // Assert
+      const collectionName = context.getCollectionName(fixturesPath)
+      expect(fakeDb.isHybridCollection(collectionName)).toBe(true)
+    })
+  })
 
-    describe('Error Handling', () => {
-        it('should handle embedding provider failure gracefully', async () => {
-            // Arrange: Create new context with failing embedding
-            const failingEmbedding = new FakeEmbedding(128);
-            failingEmbedding.injectFailure();
+  describe('error Handling', () => {
+    it('should handle embedding provider failure gracefully', async () => {
+      // Arrange: Create new context with failing embedding
+      const failingEmbedding = new FakeEmbedding(128)
+      failingEmbedding.injectFailure()
 
-            const failingContext = new TestContextBuilder()
-                .withEmbedding(failingEmbedding)
-                .withVectorDatabase(fakeDb)
-                .build();
+      const failingContext = new TestContextBuilder()
+        .withEmbedding(failingEmbedding)
+        .withVectorDatabase(fakeDb)
+        .build()
 
-            // Act: The context logs errors but continues
-            const result = await failingContext.indexCodebase(fixturesPath);
+      // Act: The context logs errors but continues
+      const result = await failingContext.indexCodebase(fixturesPath)
 
-            // Assert: Indexing completes but with potential partial results
-            expect(result).toBeDefined();
-            expect(result.status).toBeDefined();
-        });
+      // Assert: Indexing completes but with potential partial results
+      expect(result).toBeDefined()
+      expect(result.status).toBeDefined()
+    })
 
-        it('should handle vector database failure', async () => {
-            // Arrange: Inject failure into vector database
-            fakeDb.injectFailure();
+    it('should handle vector database failure', async () => {
+      // Arrange: Inject failure into vector database
+      fakeDb.injectFailure()
 
-            // Act & Assert
-            await expect(context.indexCodebase(fixturesPath)).rejects.toThrow();
-        });
+      // Act & Assert
+      await expect(context.indexCodebase(fixturesPath)).rejects.toThrow()
+    })
 
-        it('should handle non-existent codebase path', async () => {
-            // Arrange
-            const nonExistentPath = '/path/that/does/not/exist';
+    it('should handle non-existent codebase path', async () => {
+      // Arrange
+      const nonExistentPath = '/path/that/does/not/exist'
 
-            // Act & Assert
-            await expect(context.indexCodebase(nonExistentPath)).rejects.toThrow();
-        });
-    });
+      // Act & Assert
+      await expect(context.indexCodebase(nonExistentPath)).rejects.toThrow()
+    })
+  })
 
-    describe('Force Reindex', () => {
-        it('should recreate collection when force reindex is true', async () => {
-            // Arrange: Index once
-            await context.indexCodebase(fixturesPath);
+  describe('force Reindex', () => {
+    it('should recreate collection when force reindex is true', async () => {
+      // Arrange: Index once
+      await context.indexCodebase(fixturesPath)
 
-            const collectionName = context.getCollectionName(fixturesPath);
-            const firstDocCount = fakeDb.getCollectionDocumentCount(collectionName);
+      const collectionName = context.getCollectionName(fixturesPath)
+      const firstDocCount = fakeDb.getCollectionDocumentCount(collectionName)
 
-            // Act: Force reindex
-            await context.indexCodebase(fixturesPath, undefined, true);
+      // Act: Force reindex
+      await context.indexCodebase(fixturesPath, undefined, true)
 
-            // Assert: Collection still exists with documents
-            expect(await fakeDb.hasCollection(collectionName)).toBe(true);
-            const secondDocCount = fakeDb.getCollectionDocumentCount(collectionName);
-            expect(secondDocCount).toBeGreaterThan(0);
-        });
-    });
+      // Assert: Collection still exists with documents
+      expect(await fakeDb.hasCollection(collectionName)).toBe(true)
+      const secondDocCount = fakeDb.getCollectionDocumentCount(collectionName)
+      expect(secondDocCount).toBeGreaterThan(0)
+    })
+  })
 
-    describe('Chunk Metadata', () => {
-        it('should include correct file metadata in chunks', async () => {
-            // Act
-            await context.indexCodebase(fixturesPath);
+  describe('chunk Metadata', () => {
+    it('should include correct file metadata in chunks', async () => {
+      // Act
+      await context.indexCodebase(fixturesPath)
 
-            // Assert
-            const collectionName = context.getCollectionName(fixturesPath);
-            const documents = fakeDb.getStoredDocuments(collectionName);
+      // Assert
+      const collectionName = context.getCollectionName(fixturesPath)
+      const documents = fakeDb.getStoredDocuments(collectionName)
 
-            // Find a document from user-service.ts
-            const userServiceDoc = documents.find(doc =>
-                doc.relativePath.includes('user-service.ts')
-            );
+      // Find a document from user-service.ts
+      const userServiceDoc = documents.find((doc) =>
+        doc.relativePath.includes('user-service.ts'),
+      )
 
-            expect(userServiceDoc).toBeDefined();
-            expect(userServiceDoc!.fileExtension).toBe('.ts');
-            expect(userServiceDoc!.startLine).toBeGreaterThan(0);
-            expect(userServiceDoc!.endLine).toBeGreaterThanOrEqual(userServiceDoc!.startLine);
-            expect(userServiceDoc!.content.length).toBeGreaterThan(0);
-        });
+      expect(userServiceDoc).toBeDefined()
+      expect(userServiceDoc!.fileExtension).toBe('.ts')
+      expect(userServiceDoc!.startLine).toBeGreaterThan(0)
+      expect(userServiceDoc!.endLine).toBeGreaterThanOrEqual(userServiceDoc!.startLine)
+      expect(userServiceDoc!.content.length).toBeGreaterThan(0)
+    })
 
-        it('should generate unique IDs for each chunk', async () => {
-            // Act
-            await context.indexCodebase(fixturesPath);
+    it('should generate unique IDs for each chunk', async () => {
+      // Act
+      await context.indexCodebase(fixturesPath)
 
-            // Assert
-            const collectionName = context.getCollectionName(fixturesPath);
-            const documents = fakeDb.getStoredDocuments(collectionName);
+      // Assert
+      const collectionName = context.getCollectionName(fixturesPath)
+      const documents = fakeDb.getStoredDocuments(collectionName)
 
-            const ids = documents.map(doc => doc.id);
-            const uniqueIds = new Set(ids);
+      const ids = documents.map((doc) => doc.id)
+      const uniqueIds = new Set(ids)
 
-            expect(uniqueIds.size).toBe(ids.length); // All IDs are unique
-        });
-    });
-});
+      expect(uniqueIds.size).toBe(ids.length) // All IDs are unique
+    })
+  })
+})
