@@ -1,4 +1,4 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
@@ -19,4 +19,4 @@ export default defineConfig({
       ],
     },
   },
-});
+})
