@@ -1,194 +1,195 @@
-import { envManager } from "@pleaseai/context-please-core";
+import { envManager } from '@pleaseai/context-please-core'
 
 export interface ContextMcpConfig {
-    name: string;
-    version: string;
-    // Embedding provider configuration
-    embeddingProvider: 'OpenAI' | 'VoyageAI' | 'Gemini' | 'Ollama';
-    embeddingModel: string;
-    // Provider-specific API keys
-    openaiApiKey?: string;
-    openaiBaseUrl?: string;
-    voyageaiApiKey?: string;
-    geminiApiKey?: string;
-    geminiBaseUrl?: string;
-    // Ollama configuration
-    ollamaModel?: string;
-    ollamaHost?: string;
-    // Vector database configuration
-    vectorDbType?: 'milvus' | 'qdrant'; // Vector database type (default: milvus)
-    milvusAddress?: string; // Optional, can be auto-resolved from token
-    milvusToken?: string;
-    qdrantUrl?: string; // Qdrant URL (e.g., http://localhost:6333 or cloud URL)
-    qdrantApiKey?: string; // Qdrant API key (optional for self-hosted)
+  name: string
+  version: string
+  // Embedding provider configuration
+  embeddingProvider: 'OpenAI' | 'VoyageAI' | 'Gemini' | 'Ollama'
+  embeddingModel: string
+  // Provider-specific API keys
+  openaiApiKey?: string
+  openaiBaseUrl?: string
+  voyageaiApiKey?: string
+  geminiApiKey?: string
+  geminiBaseUrl?: string
+  // Ollama configuration
+  ollamaModel?: string
+  ollamaHost?: string
+  // Vector database configuration
+  vectorDbType?: 'milvus' | 'qdrant' // Vector database type (default: milvus)
+  milvusAddress?: string // Optional, can be auto-resolved from token
+  milvusToken?: string
+  qdrantUrl?: string // Qdrant URL (e.g., http://localhost:6333 or cloud URL)
+  qdrantApiKey?: string // Qdrant API key (optional for self-hosted)
 }
 
 // Legacy format (v1) - for backward compatibility
 export interface CodebaseSnapshotV1 {
-    indexedCodebases: string[];
-    indexingCodebases: string[] | Record<string, number>;  // Array (legacy) or Map of codebase path to progress percentage
-    lastUpdated: string;
+  indexedCodebases: string[]
+  indexingCodebases: string[] | Record<string, number> // Array (legacy) or Map of codebase path to progress percentage
+  lastUpdated: string
 }
 
 // New format (v2) - structured with codebase information
 
 // Base interface for common fields
 interface CodebaseInfoBase {
-    lastUpdated: string;
+  lastUpdated: string
 }
 
 // Indexing state - when indexing is in progress
 export interface CodebaseInfoIndexing extends CodebaseInfoBase {
-    status: 'indexing';
-    indexingPercentage: number;  // Current progress percentage
+  status: 'indexing'
+  indexingPercentage: number // Current progress percentage
 }
 
 // Indexed state - when indexing completed successfully
 export interface CodebaseInfoIndexed extends CodebaseInfoBase {
-    status: 'indexed';
-    indexedFiles: number;        // Number of files indexed
-    totalChunks: number;         // Total number of chunks generated
-    indexStatus: 'completed' | 'limit_reached';  // Status from indexing result
+  status: 'indexed'
+  indexedFiles: number // Number of files indexed
+  totalChunks: number // Total number of chunks generated
+  indexStatus: 'completed' | 'limit_reached' // Status from indexing result
 }
 
 // Index failed state - when indexing failed
 export interface CodebaseInfoIndexFailed extends CodebaseInfoBase {
-    status: 'indexfailed';
-    errorMessage: string;        // Error message from the failure
-    lastAttemptedPercentage?: number;  // Progress when failure occurred
+  status: 'indexfailed'
+  errorMessage: string // Error message from the failure
+  lastAttemptedPercentage?: number // Progress when failure occurred
 }
 
 // Union type for all codebase information states
-export type CodebaseInfo = CodebaseInfoIndexing | CodebaseInfoIndexed | CodebaseInfoIndexFailed;
+export type CodebaseInfo = CodebaseInfoIndexing | CodebaseInfoIndexed | CodebaseInfoIndexFailed
 
 export interface CodebaseSnapshotV2 {
-    formatVersion: 'v2';
-    codebases: Record<string, CodebaseInfo>;  // codebasePath -> CodebaseInfo
-    lastUpdated: string;
+  formatVersion: 'v2'
+  codebases: Record<string, CodebaseInfo> // codebasePath -> CodebaseInfo
+  lastUpdated: string
 }
 
 // Union type for all supported formats
-export type CodebaseSnapshot = CodebaseSnapshotV1 | CodebaseSnapshotV2;
+export type CodebaseSnapshot = CodebaseSnapshotV1 | CodebaseSnapshotV2
 
 // Helper function to get default model for each provider
 export function getDefaultModelForProvider(provider: string): string {
-    switch (provider) {
-        case 'OpenAI':
-            return 'text-embedding-3-small';
-        case 'VoyageAI':
-            return 'voyage-code-3';
-        case 'Gemini':
-            return 'gemini-embedding-001';
-        case 'Ollama':
-            return 'nomic-embed-text';
-        default:
-            return 'text-embedding-3-small';
-    }
+  switch (provider) {
+    case 'OpenAI':
+      return 'text-embedding-3-small'
+    case 'VoyageAI':
+      return 'voyage-code-3'
+    case 'Gemini':
+      return 'gemini-embedding-001'
+    case 'Ollama':
+      return 'nomic-embed-text'
+    default:
+      return 'text-embedding-3-small'
+  }
 }
 
 // Helper function to get embedding model with provider-specific environment variable priority
 export function getEmbeddingModelForProvider(provider: string): string {
-    switch (provider) {
-        case 'Ollama':
-            // For Ollama, prioritize OLLAMA_MODEL over EMBEDDING_MODEL for backward compatibility
-            const ollamaModel = envManager.get('OLLAMA_MODEL') || envManager.get('EMBEDDING_MODEL') || getDefaultModelForProvider(provider);
-            console.log(`[DEBUG] 🎯 Ollama model selection: OLLAMA_MODEL=${envManager.get('OLLAMA_MODEL') || 'NOT SET'}, EMBEDDING_MODEL=${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}, selected=${ollamaModel}`);
-            return ollamaModel;
-        case 'OpenAI':
-        case 'VoyageAI':
-        case 'Gemini':
-        default:
-            // For all other providers, use EMBEDDING_MODEL or default
-            const selectedModel = envManager.get('EMBEDDING_MODEL') || getDefaultModelForProvider(provider);
-            console.log(`[DEBUG] 🎯 ${provider} model selection: EMBEDDING_MODEL=${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}, selected=${selectedModel}`);
-            return selectedModel;
-    }
+  switch (provider) {
+    case 'Ollama':
+      // For Ollama, prioritize OLLAMA_MODEL over EMBEDDING_MODEL for backward compatibility
+      const ollamaModel = envManager.get('OLLAMA_MODEL') || envManager.get('EMBEDDING_MODEL') || getDefaultModelForProvider(provider)
+      console.log(`[DEBUG] 🎯 Ollama model selection: OLLAMA_MODEL=${envManager.get('OLLAMA_MODEL') || 'NOT SET'}, EMBEDDING_MODEL=${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}, selected=${ollamaModel}`)
+      return ollamaModel
+    case 'OpenAI':
+    case 'VoyageAI':
+    case 'Gemini':
+    default:
+      // For all other providers, use EMBEDDING_MODEL or default
+      const selectedModel = envManager.get('EMBEDDING_MODEL') || getDefaultModelForProvider(provider)
+      console.log(`[DEBUG] 🎯 ${provider} model selection: EMBEDDING_MODEL=${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}, selected=${selectedModel}`)
+      return selectedModel
+  }
 }
 
 export function createMcpConfig(): ContextMcpConfig {
-    // Debug: Print all environment variables related to Context
-    console.log(`[DEBUG] 🔍 Environment Variables Debug:`);
-    console.log(`[DEBUG]   EMBEDDING_PROVIDER: ${envManager.get('EMBEDDING_PROVIDER') || 'NOT SET'}`);
-    console.log(`[DEBUG]   EMBEDDING_MODEL: ${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}`);
-    console.log(`[DEBUG]   OLLAMA_MODEL: ${envManager.get('OLLAMA_MODEL') || 'NOT SET'}`);
-    console.log(`[DEBUG]   GEMINI_API_KEY: ${envManager.get('GEMINI_API_KEY') ? 'SET (length: ' + envManager.get('GEMINI_API_KEY')!.length + ')' : 'NOT SET'}`);
-    console.log(`[DEBUG]   OPENAI_API_KEY: ${envManager.get('OPENAI_API_KEY') ? 'SET (length: ' + envManager.get('OPENAI_API_KEY')!.length + ')' : 'NOT SET'}`);
-    console.log(`[DEBUG]   VECTOR_DB_TYPE: ${envManager.get('VECTOR_DB_TYPE') || 'NOT SET'}`);
-    console.log(`[DEBUG]   MILVUS_ADDRESS: ${envManager.get('MILVUS_ADDRESS') || 'NOT SET'}`);
-    console.log(`[DEBUG]   QDRANT_URL: ${envManager.get('QDRANT_URL') || 'NOT SET'}`);
-    console.log(`[DEBUG]   NODE_ENV: ${envManager.get('NODE_ENV') || 'NOT SET'}`);
+  // Debug: Print all environment variables related to Context
+  console.log(`[DEBUG] 🔍 Environment Variables Debug:`)
+  console.log(`[DEBUG]   EMBEDDING_PROVIDER: ${envManager.get('EMBEDDING_PROVIDER') || 'NOT SET'}`)
+  console.log(`[DEBUG]   EMBEDDING_MODEL: ${envManager.get('EMBEDDING_MODEL') || 'NOT SET'}`)
+  console.log(`[DEBUG]   OLLAMA_MODEL: ${envManager.get('OLLAMA_MODEL') || 'NOT SET'}`)
+  console.log(`[DEBUG]   GEMINI_API_KEY: ${envManager.get('GEMINI_API_KEY') ? `SET (length: ${envManager.get('GEMINI_API_KEY')!.length})` : 'NOT SET'}`)
+  console.log(`[DEBUG]   OPENAI_API_KEY: ${envManager.get('OPENAI_API_KEY') ? `SET (length: ${envManager.get('OPENAI_API_KEY')!.length})` : 'NOT SET'}`)
+  console.log(`[DEBUG]   VECTOR_DB_TYPE: ${envManager.get('VECTOR_DB_TYPE') || 'NOT SET'}`)
+  console.log(`[DEBUG]   MILVUS_ADDRESS: ${envManager.get('MILVUS_ADDRESS') || 'NOT SET'}`)
+  console.log(`[DEBUG]   QDRANT_URL: ${envManager.get('QDRANT_URL') || 'NOT SET'}`)
+  console.log(`[DEBUG]   NODE_ENV: ${envManager.get('NODE_ENV') || 'NOT SET'}`)
 
-    const config: ContextMcpConfig = {
-        name: envManager.get('MCP_SERVER_NAME') || "Context MCP Server",
-        version: envManager.get('MCP_SERVER_VERSION') || "1.0.0",
-        // Embedding provider configuration
-        embeddingProvider: (envManager.get('EMBEDDING_PROVIDER') as 'OpenAI' | 'VoyageAI' | 'Gemini' | 'Ollama') || 'OpenAI',
-        embeddingModel: getEmbeddingModelForProvider(envManager.get('EMBEDDING_PROVIDER') || 'OpenAI'),
-        // Provider-specific API keys
-        openaiApiKey: envManager.get('OPENAI_API_KEY'),
-        openaiBaseUrl: envManager.get('OPENAI_BASE_URL'),
-        voyageaiApiKey: envManager.get('VOYAGEAI_API_KEY'),
-        geminiApiKey: envManager.get('GEMINI_API_KEY'),
-        geminiBaseUrl: envManager.get('GEMINI_BASE_URL'),
-        // Ollama configuration
-        ollamaModel: envManager.get('OLLAMA_MODEL'),
-        ollamaHost: envManager.get('OLLAMA_HOST'),
-        // Vector database configuration
-        vectorDbType: (envManager.get('VECTOR_DB_TYPE') as 'milvus' | 'qdrant') || 'milvus',
-        milvusAddress: envManager.get('MILVUS_ADDRESS'), // Optional, can be resolved from token
-        milvusToken: envManager.get('MILVUS_TOKEN'),
-        qdrantUrl: envManager.get('QDRANT_URL'),
-        qdrantApiKey: envManager.get('QDRANT_API_KEY')
-    };
+  const config: ContextMcpConfig = {
+    name: envManager.get('MCP_SERVER_NAME') || 'Context MCP Server',
+    version: envManager.get('MCP_SERVER_VERSION') || '1.0.0',
+    // Embedding provider configuration
+    embeddingProvider: (envManager.get('EMBEDDING_PROVIDER') as 'OpenAI' | 'VoyageAI' | 'Gemini' | 'Ollama') || 'OpenAI',
+    embeddingModel: getEmbeddingModelForProvider(envManager.get('EMBEDDING_PROVIDER') || 'OpenAI'),
+    // Provider-specific API keys
+    openaiApiKey: envManager.get('OPENAI_API_KEY'),
+    openaiBaseUrl: envManager.get('OPENAI_BASE_URL'),
+    voyageaiApiKey: envManager.get('VOYAGEAI_API_KEY'),
+    geminiApiKey: envManager.get('GEMINI_API_KEY'),
+    geminiBaseUrl: envManager.get('GEMINI_BASE_URL'),
+    // Ollama configuration
+    ollamaModel: envManager.get('OLLAMA_MODEL'),
+    ollamaHost: envManager.get('OLLAMA_HOST'),
+    // Vector database configuration
+    vectorDbType: (envManager.get('VECTOR_DB_TYPE') as 'milvus' | 'qdrant') || 'milvus',
+    milvusAddress: envManager.get('MILVUS_ADDRESS'), // Optional, can be resolved from token
+    milvusToken: envManager.get('MILVUS_TOKEN'),
+    qdrantUrl: envManager.get('QDRANT_URL'),
+    qdrantApiKey: envManager.get('QDRANT_API_KEY'),
+  }
 
-    return config;
+  return config
 }
 
 export function logConfigurationSummary(config: ContextMcpConfig): void {
-    // Log configuration summary before starting server
-    console.log(`[MCP] 🚀 Starting Context MCP Server`);
-    console.log(`[MCP] Configuration Summary:`);
-    console.log(`[MCP]   Server: ${config.name} v${config.version}`);
-    console.log(`[MCP]   Embedding Provider: ${config.embeddingProvider}`);
-    console.log(`[MCP]   Embedding Model: ${config.embeddingModel}`);
-    console.log(`[MCP]   Vector Database: ${config.vectorDbType || 'milvus'}`);
+  // Log configuration summary before starting server
+  console.log(`[MCP] 🚀 Starting Context MCP Server`)
+  console.log(`[MCP] Configuration Summary:`)
+  console.log(`[MCP]   Server: ${config.name} v${config.version}`)
+  console.log(`[MCP]   Embedding Provider: ${config.embeddingProvider}`)
+  console.log(`[MCP]   Embedding Model: ${config.embeddingModel}`)
+  console.log(`[MCP]   Vector Database: ${config.vectorDbType || 'milvus'}`)
 
-    // Log vector database specific configuration
-    if (config.vectorDbType === 'qdrant') {
-        console.log(`[MCP]   Qdrant URL: ${config.qdrantUrl || '[Not configured]'}`);
-        console.log(`[MCP]   Qdrant API Key: ${config.qdrantApiKey ? '✅ Configured' : '❌ Not configured'}`);
-    } else {
-        console.log(`[MCP]   Milvus Address: ${config.milvusAddress || (config.milvusToken ? '[Auto-resolve from token]' : '[Not configured]')}`);
-    }
+  // Log vector database specific configuration
+  if (config.vectorDbType === 'qdrant') {
+    console.log(`[MCP]   Qdrant URL: ${config.qdrantUrl || '[Not configured]'}`)
+    console.log(`[MCP]   Qdrant API Key: ${config.qdrantApiKey ? '✅ Configured' : '❌ Not configured'}`)
+  }
+  else {
+    console.log(`[MCP]   Milvus Address: ${config.milvusAddress || (config.milvusToken ? '[Auto-resolve from token]' : '[Not configured]')}`)
+  }
 
-    // Log provider-specific configuration without exposing sensitive data
-    switch (config.embeddingProvider) {
-        case 'OpenAI':
-            console.log(`[MCP]   OpenAI API Key: ${config.openaiApiKey ? '✅ Configured' : '❌ Missing'}`);
-            if (config.openaiBaseUrl) {
-                console.log(`[MCP]   OpenAI Base URL: ${config.openaiBaseUrl}`);
-            }
-            break;
-        case 'VoyageAI':
-            console.log(`[MCP]   VoyageAI API Key: ${config.voyageaiApiKey ? '✅ Configured' : '❌ Missing'}`);
-            break;
-        case 'Gemini':
-            console.log(`[MCP]   Gemini API Key: ${config.geminiApiKey ? '✅ Configured' : '❌ Missing'}`);
-            if (config.geminiBaseUrl) {
-                console.log(`[MCP]   Gemini Base URL: ${config.geminiBaseUrl}`);
-            }
-            break;
-        case 'Ollama':
-            console.log(`[MCP]   Ollama Host: ${config.ollamaHost || 'http://127.0.0.1:11434'}`);
-            console.log(`[MCP]   Ollama Model: ${config.embeddingModel}`);
-            break;
-    }
+  // Log provider-specific configuration without exposing sensitive data
+  switch (config.embeddingProvider) {
+    case 'OpenAI':
+      console.log(`[MCP]   OpenAI API Key: ${config.openaiApiKey ? '✅ Configured' : '❌ Missing'}`)
+      if (config.openaiBaseUrl) {
+        console.log(`[MCP]   OpenAI Base URL: ${config.openaiBaseUrl}`)
+      }
+      break
+    case 'VoyageAI':
+      console.log(`[MCP]   VoyageAI API Key: ${config.voyageaiApiKey ? '✅ Configured' : '❌ Missing'}`)
+      break
+    case 'Gemini':
+      console.log(`[MCP]   Gemini API Key: ${config.geminiApiKey ? '✅ Configured' : '❌ Missing'}`)
+      if (config.geminiBaseUrl) {
+        console.log(`[MCP]   Gemini Base URL: ${config.geminiBaseUrl}`)
+      }
+      break
+    case 'Ollama':
+      console.log(`[MCP]   Ollama Host: ${config.ollamaHost || 'http://127.0.0.1:11434'}`)
+      console.log(`[MCP]   Ollama Model: ${config.embeddingModel}`)
+      break
+  }
 
-    console.log(`[MCP] 🔧 Initializing server components...`);
+  console.log(`[MCP] 🔧 Initializing server components...`)
 }
 
 export function showHelpMessage(): void {
-    console.log(`
+  console.log(`
 Context MCP Server
 
 Usage: npx @pleaseai/context-please-mcp@latest [options]
@@ -246,5 +247,5 @@ Examples:
 
   # Start MCP server with Qdrant Cloud
   OPENAI_API_KEY=sk-xxx VECTOR_DB_TYPE=qdrant QDRANT_URL=https://your-cluster.qdrant.io QDRANT_API_KEY=your-api-key npx @pleaseai/context-please-mcp@latest
-        `);
-} 
+        `)
+}
