@@ -2,115 +2,115 @@
 
 // CRITICAL: Redirect console outputs to stderr IMMEDIATELY to avoid interfering with MCP JSON protocol
 // Only MCP protocol messages should go to stdout
-const originalConsoleLog = console.log;
-const originalConsoleWarn = console.warn;
-
-console.log = (...args: any[]) => {
-    process.stderr.write('[LOG] ' + args.join(' ') + '\n');
-};
-
-console.warn = (...args: any[]) => {
-    process.stderr.write('[WARN] ' + args.join(' ') + '\n');
-};
-
 // console.error already goes to stderr by default
 
-import { Server } from "@modelcontextprotocol/sdk/server/index.js";
-import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
+import { Server } from '@modelcontextprotocol/sdk/server/index.js'
+import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js'
 import {
-    ListToolsRequestSchema,
-    CallToolRequestSchema
-} from "@modelcontextprotocol/sdk/types.js";
-import { Context } from "@pleaseai/context-please-core";
-import { MilvusVectorDatabase, QdrantVectorDatabase, VectorDatabase } from "@pleaseai/context-please-core";
+  CallToolRequestSchema,
+  ListToolsRequestSchema,
+} from '@modelcontextprotocol/sdk/types.js'
+import { Context, MilvusVectorDatabase, QdrantVectorDatabase, VectorDatabase } from '@pleaseai/context-please-core'
 
 // Import our modular components
-import { createMcpConfig, logConfigurationSummary, showHelpMessage, ContextMcpConfig } from "./config.js";
-import { createEmbeddingInstance, logEmbeddingProviderInfo } from "./embedding.js";
-import { SnapshotManager } from "./snapshot.js";
-import { SyncManager } from "./sync.js";
-import { ToolHandlers } from "./handlers.js";
+import { ContextMcpConfig, createMcpConfig, logConfigurationSummary, showHelpMessage } from './config.js'
+import { createEmbeddingInstance, logEmbeddingProviderInfo } from './embedding.js'
+import { ToolHandlers } from './handlers.js'
+import { SnapshotManager } from './snapshot.js'
+import { SyncManager } from './sync.js'
+
+const originalConsoleLog = console.log
+const originalConsoleWarn = console.warn
+
+console.log = (...args: any[]) => {
+  process.stderr.write(`[LOG] ${args.join(' ')}\n`)
+}
+
+console.warn = (...args: any[]) => {
+  process.stderr.write(`[WARN] ${args.join(' ')}\n`)
+}
 
 class ContextMcpServer {
-    private server: Server;
-    private context: Context;
-    private snapshotManager: SnapshotManager;
-    private syncManager: SyncManager;
-    private toolHandlers: ToolHandlers;
+  private server: Server
+  private context: Context
+  private snapshotManager: SnapshotManager
+  private syncManager: SyncManager
+  private toolHandlers: ToolHandlers
 
-    constructor(config: ContextMcpConfig) {
-        // Initialize MCP server
-        this.server = new Server(
-            {
-                name: config.name,
-                version: config.version
-            },
-            {
-                capabilities: {
-                    tools: {}
-                }
-            }
-        );
+  constructor(config: ContextMcpConfig) {
+    // Initialize MCP server
+    this.server = new Server(
+      {
+        name: config.name,
+        version: config.version,
+      },
+      {
+        capabilities: {
+          tools: {},
+        },
+      },
+    )
 
-        // Initialize embedding provider
-        console.log(`[EMBEDDING] Initializing embedding provider: ${config.embeddingProvider}`);
-        console.log(`[EMBEDDING] Using model: ${config.embeddingModel}`);
+    // Initialize embedding provider
+    console.log(`[EMBEDDING] Initializing embedding provider: ${config.embeddingProvider}`)
+    console.log(`[EMBEDDING] Using model: ${config.embeddingModel}`)
 
-        const embedding = createEmbeddingInstance(config);
-        logEmbeddingProviderInfo(config, embedding);
+    const embedding = createEmbeddingInstance(config)
+    logEmbeddingProviderInfo(config, embedding)
 
-        // Initialize vector database based on configuration
-        console.log(`[VECTORDB] Initializing vector database: ${config.vectorDbType || 'milvus'}`);
+    // Initialize vector database based on configuration
+    console.log(`[VECTORDB] Initializing vector database: ${config.vectorDbType || 'milvus'}`)
 
-        let vectorDatabase: VectorDatabase;
+    let vectorDatabase: VectorDatabase
 
-        if (config.vectorDbType === 'qdrant') {
-            // Parse Qdrant URL to get address for gRPC
-            const qdrantUrl = config.qdrantUrl || 'http://localhost:6333';
-            const url = new URL(qdrantUrl.startsWith('http') ? qdrantUrl : `http://${qdrantUrl}`);
+    if (config.vectorDbType === 'qdrant') {
+      // Parse Qdrant URL to get address for gRPC
+      const qdrantUrl = config.qdrantUrl || 'http://localhost:6333'
+      const url = new URL(qdrantUrl.startsWith('http') ? qdrantUrl : `http://${qdrantUrl}`)
 
-            // For Qdrant gRPC, we need host:port format.
-            // Auto-convert default REST port (6333) to default gRPC port (6334).
-            let grpcPort = url.port || '6334';
-            if (grpcPort === '6333') {
-                console.log('[VECTORDB] Qdrant REST port 6333 detected, switching to gRPC port 6334.');
-                grpcPort = '6334';
-            }
-            const grpcAddress = `${url.hostname}:${grpcPort}`;
+      // For Qdrant gRPC, we need host:port format.
+      // Auto-convert default REST port (6333) to default gRPC port (6334).
+      let grpcPort = url.port || '6334'
+      if (grpcPort === '6333') {
+        console.log('[VECTORDB] Qdrant REST port 6333 detected, switching to gRPC port 6334.')
+        grpcPort = '6334'
+      }
+      const grpcAddress = `${url.hostname}:${grpcPort}`
 
-            console.log(`[VECTORDB] Qdrant gRPC address: ${grpcAddress}`);
+      console.log(`[VECTORDB] Qdrant gRPC address: ${grpcAddress}`)
 
-            vectorDatabase = new QdrantVectorDatabase({
-                address: grpcAddress,
-                ...(config.qdrantApiKey && { apiKey: config.qdrantApiKey })
-            });
-        } else {
-            // Default to Milvus
-            vectorDatabase = new MilvusVectorDatabase({
-                address: config.milvusAddress,
-                ...(config.milvusToken && { token: config.milvusToken })
-            });
-        }
-
-        // Initialize Claude Context
-        this.context = new Context({
-            embedding,
-            vectorDatabase
-        });
-
-        // Initialize managers
-        this.snapshotManager = new SnapshotManager();
-        this.syncManager = new SyncManager(this.context, this.snapshotManager);
-        this.toolHandlers = new ToolHandlers(this.context, this.snapshotManager);
-
-        // Load existing codebase snapshot on startup
-        this.snapshotManager.loadCodebaseSnapshot();
-
-        this.setupTools();
+      vectorDatabase = new QdrantVectorDatabase({
+        address: grpcAddress,
+        ...(config.qdrantApiKey && { apiKey: config.qdrantApiKey }),
+      })
+    }
+    else {
+      // Default to Milvus
+      vectorDatabase = new MilvusVectorDatabase({
+        address: config.milvusAddress,
+        ...(config.milvusToken && { token: config.milvusToken }),
+      })
     }
 
-    private setupTools() {
-        const index_description = `
+    // Initialize Claude Context
+    this.context = new Context({
+      embedding,
+      vectorDatabase,
+    })
+
+    // Initialize managers
+    this.snapshotManager = new SnapshotManager()
+    this.syncManager = new SyncManager(this.context, this.snapshotManager)
+    this.toolHandlers = new ToolHandlers(this.context, this.snapshotManager)
+
+    // Load existing codebase snapshot on startup
+    this.snapshotManager.loadCodebaseSnapshot()
+
+    this.setupTools()
+  }
+
+  private setupTools() {
+    const index_description = `
 Index a codebase directory to enable semantic search using a configurable code splitter.
 
 ⚠️ **IMPORTANT**:
@@ -119,10 +119,9 @@ Index a codebase directory to enable semantic search using a configurable code s
 ✨ **Usage Guidance**:
 - This tool is typically used when search fails due to an unindexed codebase.
 - If indexing is attempted on an already indexed path, and a conflict is detected, you MUST prompt the user to confirm whether to proceed with a force index (i.e., re-indexing and overwriting the previous index).
-`;
+`
 
-
-        const search_description = `
+    const search_description = `
 Search the indexed codebase using natural language queries within a specified absolute path.
 
 ⚠️ **IMPORTANT**:
@@ -141,187 +140,187 @@ This tool is versatile and can be used before completing various tasks to retrie
 ✨ **Usage Guidance**:
 - If the codebase is not indexed, this tool will return a clear error message indicating that indexing is required first.
 - You can then use the index_codebase tool to index the codebase before searching again.
-`;
+`
 
-        // Define available tools
-        this.server.setRequestHandler(ListToolsRequestSchema, async () => {
-            return {
-                tools: [
-                    {
-                        name: "index_codebase",
-                        description: index_description,
-                        inputSchema: {
-                            type: "object",
-                            properties: {
-                                path: {
-                                    type: "string",
-                                    description: `ABSOLUTE path to the codebase directory to index.`
-                                },
-                                force: {
-                                    type: "boolean",
-                                    description: "Force re-indexing even if already indexed",
-                                    default: false
-                                },
-                                splitter: {
-                                    type: "string",
-                                    description: "Code splitter to use: 'ast' for syntax-aware splitting with automatic fallback, 'langchain' for character-based splitting",
-                                    enum: ["ast", "langchain"],
-                                    default: "ast"
-                                },
-                                customExtensions: {
-                                    type: "array",
-                                    items: {
-                                        type: "string"
-                                    },
-                                    description: "Optional: Additional file extensions to include beyond defaults (e.g., ['.vue', '.svelte', '.astro']). Extensions should include the dot prefix or will be automatically added",
-                                    default: []
-                                },
-                                ignorePatterns: {
-                                    type: "array",
-                                    items: {
-                                        type: "string"
-                                    },
-                                    description: "Optional: Additional ignore patterns to exclude specific files/directories beyond defaults. Only include this parameter if the user explicitly requests custom ignore patterns (e.g., ['static/**', '*.tmp', 'private/**'])",
-                                    default: []
-                                }
-                            },
-                            required: ["path"]
-                        }
-                    },
-                    {
-                        name: "search_code",
-                        description: search_description,
-                        inputSchema: {
-                            type: "object",
-                            properties: {
-                                path: {
-                                    type: "string",
-                                    description: `ABSOLUTE path to the codebase directory to search in.`
-                                },
-                                query: {
-                                    type: "string",
-                                    description: "Natural language query to search for in the codebase"
-                                },
-                                limit: {
-                                    type: "number",
-                                    description: "Maximum number of results to return",
-                                    default: 10,
-                                    maximum: 50
-                                },
-                                extensionFilter: {
-                                    type: "array",
-                                    items: {
-                                        type: "string"
-                                    },
-                                    description: "Optional: List of file extensions to filter results. (e.g., ['.ts','.py']).",
-                                    default: []
-                                }
-                            },
-                            required: ["path", "query"]
-                        }
-                    },
-                    {
-                        name: "clear_index",
-                        description: `Clear the search index. IMPORTANT: You MUST provide an absolute path.`,
-                        inputSchema: {
-                            type: "object",
-                            properties: {
-                                path: {
-                                    type: "string",
-                                    description: `ABSOLUTE path to the codebase directory to clear.`
-                                }
-                            },
-                            required: ["path"]
-                        }
-                    },
-                    {
-                        name: "get_indexing_status",
-                        description: `Get the current indexing status of a codebase. Shows progress percentage for actively indexing codebases and completion status for indexed codebases.`,
-                        inputSchema: {
-                            type: "object",
-                            properties: {
-                                path: {
-                                    type: "string",
-                                    description: `ABSOLUTE path to the codebase directory to check status for.`
-                                }
-                            },
-                            required: ["path"]
-                        }
-                    },
-                ]
-            };
-        });
+    // Define available tools
+    this.server.setRequestHandler(ListToolsRequestSchema, async () => {
+      return {
+        tools: [
+          {
+            name: 'index_codebase',
+            description: index_description,
+            inputSchema: {
+              type: 'object',
+              properties: {
+                path: {
+                  type: 'string',
+                  description: `ABSOLUTE path to the codebase directory to index.`,
+                },
+                force: {
+                  type: 'boolean',
+                  description: 'Force re-indexing even if already indexed',
+                  default: false,
+                },
+                splitter: {
+                  type: 'string',
+                  description: 'Code splitter to use: \'ast\' for syntax-aware splitting with automatic fallback, \'langchain\' for character-based splitting',
+                  enum: ['ast', 'langchain'],
+                  default: 'ast',
+                },
+                customExtensions: {
+                  type: 'array',
+                  items: {
+                    type: 'string',
+                  },
+                  description: 'Optional: Additional file extensions to include beyond defaults (e.g., [\'.vue\', \'.svelte\', \'.astro\']). Extensions should include the dot prefix or will be automatically added',
+                  default: [],
+                },
+                ignorePatterns: {
+                  type: 'array',
+                  items: {
+                    type: 'string',
+                  },
+                  description: 'Optional: Additional ignore patterns to exclude specific files/directories beyond defaults. Only include this parameter if the user explicitly requests custom ignore patterns (e.g., [\'static/**\', \'*.tmp\', \'private/**\'])',
+                  default: [],
+                },
+              },
+              required: ['path'],
+            },
+          },
+          {
+            name: 'search_code',
+            description: search_description,
+            inputSchema: {
+              type: 'object',
+              properties: {
+                path: {
+                  type: 'string',
+                  description: `ABSOLUTE path to the codebase directory to search in.`,
+                },
+                query: {
+                  type: 'string',
+                  description: 'Natural language query to search for in the codebase',
+                },
+                limit: {
+                  type: 'number',
+                  description: 'Maximum number of results to return',
+                  default: 10,
+                  maximum: 50,
+                },
+                extensionFilter: {
+                  type: 'array',
+                  items: {
+                    type: 'string',
+                  },
+                  description: 'Optional: List of file extensions to filter results. (e.g., [\'.ts\',\'.py\']).',
+                  default: [],
+                },
+              },
+              required: ['path', 'query'],
+            },
+          },
+          {
+            name: 'clear_index',
+            description: `Clear the search index. IMPORTANT: You MUST provide an absolute path.`,
+            inputSchema: {
+              type: 'object',
+              properties: {
+                path: {
+                  type: 'string',
+                  description: `ABSOLUTE path to the codebase directory to clear.`,
+                },
+              },
+              required: ['path'],
+            },
+          },
+          {
+            name: 'get_indexing_status',
+            description: `Get the current indexing status of a codebase. Shows progress percentage for actively indexing codebases and completion status for indexed codebases.`,
+            inputSchema: {
+              type: 'object',
+              properties: {
+                path: {
+                  type: 'string',
+                  description: `ABSOLUTE path to the codebase directory to check status for.`,
+                },
+              },
+              required: ['path'],
+            },
+          },
+        ],
+      }
+    })
 
-        // Handle tool execution
-        this.server.setRequestHandler(CallToolRequestSchema, async (request) => {
-            const { name, arguments: args } = request.params;
+    // Handle tool execution
+    this.server.setRequestHandler(CallToolRequestSchema, async (request) => {
+      const { name, arguments: args } = request.params
 
-            switch (name) {
-                case "index_codebase":
-                    return await this.toolHandlers.handleIndexCodebase(args);
-                case "search_code":
-                    return await this.toolHandlers.handleSearchCode(args);
-                case "clear_index":
-                    return await this.toolHandlers.handleClearIndex(args);
-                case "get_indexing_status":
-                    return await this.toolHandlers.handleGetIndexingStatus(args);
+      switch (name) {
+        case 'index_codebase':
+          return await this.toolHandlers.handleIndexCodebase(args)
+        case 'search_code':
+          return await this.toolHandlers.handleSearchCode(args)
+        case 'clear_index':
+          return await this.toolHandlers.handleClearIndex(args)
+        case 'get_indexing_status':
+          return await this.toolHandlers.handleGetIndexingStatus(args)
 
-                default:
-                    throw new Error(`Unknown tool: ${name}`);
-            }
-        });
-    }
+        default:
+          throw new Error(`Unknown tool: ${name}`)
+      }
+    })
+  }
 
-    async start() {
-        console.log('[SYNC-DEBUG] MCP server start() method called');
-        console.log('Starting Context MCP server...');
+  async start() {
+    console.log('[SYNC-DEBUG] MCP server start() method called')
+    console.log('Starting Context MCP server...')
 
-        const transport = new StdioServerTransport();
-        console.log('[SYNC-DEBUG] StdioServerTransport created, attempting server connection...');
+    const transport = new StdioServerTransport()
+    console.log('[SYNC-DEBUG] StdioServerTransport created, attempting server connection...')
 
-        await this.server.connect(transport);
-        console.log("MCP server started and listening on stdio.");
-        console.log('[SYNC-DEBUG] Server connection established successfully');
+    await this.server.connect(transport)
+    console.log('MCP server started and listening on stdio.')
+    console.log('[SYNC-DEBUG] Server connection established successfully')
 
-        // Start background sync after server is connected
-        console.log('[SYNC-DEBUG] Initializing background sync...');
-        this.syncManager.startBackgroundSync();
-        console.log('[SYNC-DEBUG] MCP server initialization complete');
-    }
+    // Start background sync after server is connected
+    console.log('[SYNC-DEBUG] Initializing background sync...')
+    this.syncManager.startBackgroundSync()
+    console.log('[SYNC-DEBUG] MCP server initialization complete')
+  }
 }
 
 // Main execution
 async function main() {
-    // Parse command line arguments
-    const args = process.argv.slice(2);
+  // Parse command line arguments
+  const args = process.argv.slice(2)
 
-    // Show help if requested
-    if (args.includes('--help') || args.includes('-h')) {
-        showHelpMessage();
-        process.exit(0);
-    }
+  // Show help if requested
+  if (args.includes('--help') || args.includes('-h')) {
+    showHelpMessage()
+    process.exit(0)
+  }
 
-    // Create configuration
-    const config = createMcpConfig();
-    logConfigurationSummary(config);
+  // Create configuration
+  const config = createMcpConfig()
+  logConfigurationSummary(config)
 
-    const server = new ContextMcpServer(config);
-    await server.start();
+  const server = new ContextMcpServer(config)
+  await server.start()
 }
 
 // Handle graceful shutdown
 process.on('SIGINT', () => {
-    console.error("Received SIGINT, shutting down gracefully...");
-    process.exit(0);
-});
+  console.error('Received SIGINT, shutting down gracefully...')
+  process.exit(0)
+})
 
 process.on('SIGTERM', () => {
-    console.error("Received SIGTERM, shutting down gracefully...");
-    process.exit(0);
-});
+  console.error('Received SIGTERM, shutting down gracefully...')
+  process.exit(0)
+})
 
 // Always start the server - this is designed to be the main entry point
 main().catch((error) => {
-    console.error("Fatal error:", error);
-    process.exit(1);
-});
+  console.error('Fatal error:', error)
+  process.exit(1)
+})
