@@ -1,9 +1,9 @@
-import { describe, it, expect, beforeEach, afterEach } from 'vitest';
-import * as path from 'path';
-import { Context, QdrantVectorDatabase } from '@pleaseai/context-please-core';
-import { ToolHandlers } from '../../src/handlers.js';
-import { FakeSnapshotManager } from '../doubles/fake-snapshot-manager.js';
-import { FakeEmbedding } from '../../../core/test/doubles/fake-embedding.js';
+import * as path from 'node:path'
+import { Context, QdrantVectorDatabase } from '@pleaseai/context-please-core'
+import { afterEach, beforeEach, describe, expect, it } from 'vitest'
+import { FakeEmbedding } from '../../../core/test/doubles/fake-embedding.js'
+import { ToolHandlers } from '../../src/handlers.js'
+import { FakeSnapshotManager } from '../doubles/fake-snapshot-manager.js'
 
 /**
  * Integration test for Qdrant-specific syncIndexedCodebasesFromCloud() issue
@@ -14,212 +14,221 @@ import { FakeEmbedding } from '../../../core/test/doubles/fake-embedding.js';
  * 3. Empty collection causes codebase to be removed from snapshot
  * 4. search_code returns "not indexed" error
  */
-describe('Qdrant Sync During Background Indexing', () => {
-    let handlers: ToolHandlers;
-    let context: Context;
-    let snapshotManager: FakeSnapshotManager;
-    let qdrantDb: QdrantVectorDatabase;
-    let fixturesPath: string;
+describe('qdrant Sync During Background Indexing', () => {
+  let handlers: ToolHandlers
+  let context: Context
+  let snapshotManager: FakeSnapshotManager
+  let qdrantDb: QdrantVectorDatabase
+  let fixturesPath: string
 
-    // Use real Qdrant if available, otherwise skip tests
-    const qdrantUrl = process.env.QDRANT_URL || 'http://localhost:6333';
-    const skipQdrant = !process.env.QDRANT_URL && !process.env.CI;
+  // Use real Qdrant if available, otherwise skip tests
+  const qdrantUrl = process.env.QDRANT_URL || 'http://localhost:6333'
+  const skipQdrant = !process.env.QDRANT_URL && !process.env.CI
 
-    beforeEach(async () => {
-        if (skipQdrant) {
-            console.log('⏭️  Skipping Qdrant tests (QDRANT_URL not set)');
-            return;
-        }
+  beforeEach(async () => {
+    if (skipQdrant) {
+      console.log('⏭️  Skipping Qdrant tests (QDRANT_URL not set)')
+      return
+    }
 
-        // Create real Qdrant connection
-        qdrantDb = new QdrantVectorDatabase({
-            address: qdrantUrl
-        });
+    // Create real Qdrant connection
+    qdrantDb = new QdrantVectorDatabase({
+      address: qdrantUrl,
+    })
 
-        // Create Context with fake embedding and real Qdrant
-        const fakeEmbedding = new FakeEmbedding();
-        context = new Context({
-            embedding: fakeEmbedding,
-            vectorDatabase: qdrantDb
-        });
+    // Create Context with fake embedding and real Qdrant
+    const fakeEmbedding = new FakeEmbedding()
+    context = new Context({
+      embedding: fakeEmbedding,
+      vectorDatabase: qdrantDb,
+    })
 
-        // Create fake snapshot manager
-        snapshotManager = new FakeSnapshotManager();
+    // Create fake snapshot manager
+    snapshotManager = new FakeSnapshotManager()
 
-        // Create handlers
-        handlers = new ToolHandlers(context, snapshotManager);
+    // Create handlers
+    handlers = new ToolHandlers(context, snapshotManager)
 
-        // Path to test fixtures
-        fixturesPath = path.join(__dirname, '../../../core/test/fixtures/sample-codebase');
-    });
+    // Path to test fixtures
+    fixturesPath = path.join(__dirname, '../../../core/test/fixtures/sample-codebase')
+  })
 
-    afterEach(async () => {
-        if (skipQdrant) return;
+  afterEach(async () => {
+    if (skipQdrant)
+      return
 
-        // Clean up: Clear any test collections
-        try {
-            await context.clearIndex(fixturesPath);
-        } catch (e) {
-            // Ignore if collection doesn't exist
-        }
+    // Clean up: Clear any test collections
+    try {
+      await context.clearIndex(fixturesPath)
+    }
+    catch (e) {
+      // Ignore if collection doesn't exist
+    }
 
-        snapshotManager.reset();
-    });
+    snapshotManager.reset()
+  })
 
-    it('should reproduce: syncIndexedCodebasesFromCloud removes indexing codebase from snapshot', async () => {
-        if (skipQdrant) return;
+  it('should reproduce: syncIndexedCodebasesFromCloud removes indexing codebase from snapshot', async () => {
+    if (skipQdrant)
+      return
 
-        // Arrange: Simulate background indexing starting
-        // 1. Create collection (happens during indexing initialization)
-        const collectionName = context.getCollectionName(fixturesPath);
-        await qdrantDb.createCollection(collectionName, 1536, { useSparse: true });
+    // Arrange: Simulate background indexing starting
+    // 1. Create collection (happens during indexing initialization)
+    const collectionName = context.getCollectionName(fixturesPath)
+    await qdrantDb.createCollection(collectionName, 1536, { useSparse: true })
 
-        // 2. Mark as indexing in snapshot
-        snapshotManager.setCodebaseIndexing(fixturesPath, 10); // 10% progress
+    // 2. Mark as indexing in snapshot
+    snapshotManager.setCodebaseIndexing(fixturesPath, 10) // 10% progress
 
-        // 3. Collection exists but is empty (no documents inserted yet)
-        const collections = await qdrantDb.listCollections();
-        expect(collections).toContain(collectionName);
+    // 3. Collection exists but is empty (no documents inserted yet)
+    const collections = await qdrantDb.listCollections()
+    expect(collections).toContain(collectionName)
 
-        // Act: Call syncIndexedCodebasesFromCloud (happens in handleSearchCode)
-        // This simulates what happens when search_code is called during indexing
-        await (handlers as any).syncIndexedCodebasesFromCloud();
+    // Act: Call syncIndexedCodebasesFromCloud (happens in handleSearchCode)
+    // This simulates what happens when search_code is called during indexing
+    await (handlers as any).syncIndexedCodebasesFromCloud()
 
-        // Assert: Codebase should NOT be removed from snapshot
-        // (This is where the bug occurs - it gets removed because collection is empty)
-        const indexingCodebases = snapshotManager.getIndexingCodebases();
-        expect(indexingCodebases).toContain(fixturesPath);
-    });
+    // Assert: Codebase should NOT be removed from snapshot
+    // (This is where the bug occurs - it gets removed because collection is empty)
+    const indexingCodebases = snapshotManager.getIndexingCodebases()
+    expect(indexingCodebases).toContain(fixturesPath)
+  })
 
-    it('should fix: syncIndexedCodebasesFromCloud preserves indexing codebases', async () => {
-        if (skipQdrant) return;
+  it('should fix: syncIndexedCodebasesFromCloud preserves indexing codebases', async () => {
+    if (skipQdrant)
+      return
 
-        // Arrange: Same setup as above
-        const collectionName = context.getCollectionName(fixturesPath);
-        await qdrantDb.createCollection(collectionName, 1536, { useSparse: true });
-        snapshotManager.setCodebaseIndexing(fixturesPath, 25);
+    // Arrange: Same setup as above
+    const collectionName = context.getCollectionName(fixturesPath)
+    await qdrantDb.createCollection(collectionName, 1536, { useSparse: true })
+    snapshotManager.setCodebaseIndexing(fixturesPath, 25)
 
-        // Act: Call sync
-        await (handlers as any).syncIndexedCodebasesFromCloud();
+    // Act: Call sync
+    await (handlers as any).syncIndexedCodebasesFromCloud()
 
-        // Assert: With the fix, indexing codebases should be preserved
-        const indexingCodebases = snapshotManager.getIndexingCodebases();
-        expect(indexingCodebases).toContain(fixturesPath);
+    // Assert: With the fix, indexing codebases should be preserved
+    const indexingCodebases = snapshotManager.getIndexingCodebases()
+    expect(indexingCodebases).toContain(fixturesPath)
 
-        // Verify snapshot is not empty
-        const allCodebases = [...snapshotManager.getIndexedCodebases(), ...indexingCodebases];
-        expect(allCodebases.length).toBeGreaterThan(0);
-    });
+    // Verify snapshot is not empty
+    const allCodebases = [...snapshotManager.getIndexedCodebases(), ...indexingCodebases]
+    expect(allCodebases.length).toBeGreaterThan(0)
+  })
 
-    it('should allow search during indexing after sync', async () => {
-        if (skipQdrant) return;
+  it('should allow search during indexing after sync', async () => {
+    if (skipQdrant)
+      return
 
-        // Arrange: Fully index a codebase
-        await context.indexCodebase(fixturesPath);
-        snapshotManager.setCodebaseIndexing(fixturesPath, 50);
+    // Arrange: Fully index a codebase
+    await context.indexCodebase(fixturesPath)
+    snapshotManager.setCodebaseIndexing(fixturesPath, 50)
 
-        // Act 1: Sync (should preserve indexing status)
-        await (handlers as any).syncIndexedCodebasesFromCloud();
+    // Act 1: Sync (should preserve indexing status)
+    await (handlers as any).syncIndexedCodebasesFromCloud()
 
-        // Act 2: Try to search
-        const searchResult = await handlers.handleSearchCode({
-            path: fixturesPath,
-            query: 'user service',
-            limit: 5
-        });
+    // Act 2: Try to search
+    const searchResult = await handlers.handleSearchCode({
+      path: fixturesPath,
+      query: 'user service',
+      limit: 5,
+    })
 
-        // Assert: Search should work (not return "not indexed" error)
-        expect(searchResult.isError).not.toBe(true);
-        expect(searchResult.content[0].text).not.toContain('not indexed');
-        expect(searchResult.content[0].text).toContain('Indexing in Progress');
-    });
+    // Assert: Search should work (not return "not indexed" error)
+    expect(searchResult.isError).not.toBe(true)
+    expect(searchResult.content[0].text).not.toContain('not indexed')
+    expect(searchResult.content[0].text).toContain('Indexing in Progress')
+  })
 
-    it('should handle completed indexing correctly after sync', async () => {
-        if (skipQdrant) return;
+  it('should handle completed indexing correctly after sync', async () => {
+    if (skipQdrant)
+      return
 
-        // Arrange: Fully index and mark as completed
-        await context.indexCodebase(fixturesPath);
-        snapshotManager.setCodebaseIndexed(fixturesPath, {
-            indexedFiles: 3,
-            totalChunks: 28,
-            status: 'completed'
-        });
+    // Arrange: Fully index and mark as completed
+    await context.indexCodebase(fixturesPath)
+    snapshotManager.setCodebaseIndexed(fixturesPath, {
+      indexedFiles: 3,
+      totalChunks: 28,
+      status: 'completed',
+    })
 
-        // Act: Sync
-        await (handlers as any).syncIndexedCodebasesFromCloud();
+    // Act: Sync
+    await (handlers as any).syncIndexedCodebasesFromCloud()
 
-        // Assert: Indexed status should be preserved
-        const indexedCodebases = snapshotManager.getIndexedCodebases();
-        expect(indexedCodebases).toContain(fixturesPath);
+    // Assert: Indexed status should be preserved
+    const indexedCodebases = snapshotManager.getIndexedCodebases()
+    expect(indexedCodebases).toContain(fixturesPath)
 
-        // Search should work
-        const searchResult = await handlers.handleSearchCode({
-            path: fixturesPath,
-            query: 'user',
-            limit: 5
-        });
+    // Search should work
+    const searchResult = await handlers.handleSearchCode({
+      path: fixturesPath,
+      query: 'user',
+      limit: 5,
+    })
 
-        expect(searchResult.isError).not.toBe(true);
-        expect(searchResult.content[0].text).toContain('Found');
-    });
+    expect(searchResult.isError).not.toBe(true)
+    expect(searchResult.content[0].text).toContain('Found')
+  })
 
-    it('should remove truly orphaned indexed codebases from snapshot', async () => {
-        if (skipQdrant) return;
+  it('should remove truly orphaned indexed codebases from snapshot', async () => {
+    if (skipQdrant)
+      return
 
-        // Arrange: Add a codebase to snapshot that has NO collection in Qdrant
-        // This codebase is marked as "indexed" (not "indexing"), so it should be removed
-        const orphanedPath = '/tmp/orphaned-codebase-that-never-existed';
-        snapshotManager.setCodebaseIndexed(orphanedPath, {
-            indexedFiles: 10,
-            totalChunks: 50,
-            status: 'completed'
-        });
+    // Arrange: Add a codebase to snapshot that has NO collection in Qdrant
+    // This codebase is marked as "indexed" (not "indexing"), so it should be removed
+    const orphanedPath = '/tmp/orphaned-codebase-that-never-existed'
+    snapshotManager.setCodebaseIndexed(orphanedPath, {
+      indexedFiles: 10,
+      totalChunks: 50,
+      status: 'completed',
+    })
 
-        // Verify it's in snapshot before sync
-        expect(snapshotManager.getIndexedCodebases()).toContain(orphanedPath);
+    // Verify it's in snapshot before sync
+    expect(snapshotManager.getIndexedCodebases()).toContain(orphanedPath)
 
-        // Act: Sync
-        await (handlers as any).syncIndexedCodebasesFromCloud();
+    // Act: Sync
+    await (handlers as any).syncIndexedCodebasesFromCloud()
 
-        // Assert: Orphaned codebase should be removed (because it's not indexing)
-        const indexedCodebases = snapshotManager.getIndexedCodebases();
-        expect(indexedCodebases).not.toContain(orphanedPath);
-    });
+    // Assert: Orphaned codebase should be removed (because it's not indexing)
+    const indexedCodebases = snapshotManager.getIndexedCodebases()
+    expect(indexedCodebases).not.toContain(orphanedPath)
+  })
 
-    it('should NOT remove orphaned indexing codebases from snapshot', async () => {
-        if (skipQdrant) return;
+  it('should NOT remove orphaned indexing codebases from snapshot', async () => {
+    if (skipQdrant)
+      return
 
-        // Arrange: Add a codebase that is currently "indexing" but has no collection yet
-        // This simulates the race condition where sync happens before collection is created
-        const indexingOrphanPath = '/tmp/indexing-orphan-codebase';
-        snapshotManager.setCodebaseIndexing(indexingOrphanPath, 5);
+    // Arrange: Add a codebase that is currently "indexing" but has no collection yet
+    // This simulates the race condition where sync happens before collection is created
+    const indexingOrphanPath = '/tmp/indexing-orphan-codebase'
+    snapshotManager.setCodebaseIndexing(indexingOrphanPath, 5)
 
-        // Verify it's in snapshot before sync
-        expect(snapshotManager.getIndexingCodebases()).toContain(indexingOrphanPath);
+    // Verify it's in snapshot before sync
+    expect(snapshotManager.getIndexingCodebases()).toContain(indexingOrphanPath)
 
-        // Act: Sync
-        await (handlers as any).syncIndexedCodebasesFromCloud();
+    // Act: Sync
+    await (handlers as any).syncIndexedCodebasesFromCloud()
 
-        // Assert: Indexing codebase should NOT be removed (this is the fix)
-        const indexingCodebases = snapshotManager.getIndexingCodebases();
-        expect(indexingCodebases).toContain(indexingOrphanPath);
-    });
+    // Assert: Indexing codebase should NOT be removed (this is the fix)
+    const indexingCodebases = snapshotManager.getIndexingCodebases()
+    expect(indexingCodebases).toContain(indexingOrphanPath)
+  })
 
-    it('should handle empty collections during indexing gracefully', async () => {
-        if (skipQdrant) return;
+  it('should handle empty collections during indexing gracefully', async () => {
+    if (skipQdrant)
+      return
 
-        // Arrange: Create collection but don't insert any documents yet
-        const collectionName = context.getCollectionName(fixturesPath);
-        await qdrantDb.createCollection(collectionName, 1536, { useSparse: true });
-        snapshotManager.setCodebaseIndexing(fixturesPath, 0); // Just started
+    // Arrange: Create collection but don't insert any documents yet
+    const collectionName = context.getCollectionName(fixturesPath)
+    await qdrantDb.createCollection(collectionName, 1536, { useSparse: true })
+    snapshotManager.setCodebaseIndexing(fixturesPath, 0) // Just started
 
-        // Act: Multiple syncs (simulates multiple search attempts)
-        await (handlers as any).syncIndexedCodebasesFromCloud();
-        await (handlers as any).syncIndexedCodebasesFromCloud();
-        await (handlers as any).syncIndexedCodebasesFromCloud();
+    // Act: Multiple syncs (simulates multiple search attempts)
+    await (handlers as any).syncIndexedCodebasesFromCloud()
+    await (handlers as any).syncIndexedCodebasesFromCloud()
+    await (handlers as any).syncIndexedCodebasesFromCloud()
 
-        // Assert: Codebase should still be in snapshot
-        const indexingCodebases = snapshotManager.getIndexingCodebases();
-        expect(indexingCodebases).toContain(fixturesPath);
-    });
-});
+    // Assert: Codebase should still be in snapshot
+    const indexingCodebases = snapshotManager.getIndexingCodebases()
+    expect(indexingCodebases).toContain(fixturesPath)
+  })
+})
