@@ -1,4 +1,4 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
@@ -8,4 +8,4 @@ export default defineConfig({
     testTimeout: 60000, // Longer timeout for integration tests
     hookTimeout: 30000,
   },
-});
+})
