@@ -1,4 +1,4 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
@@ -9,4 +9,4 @@ export default defineConfig({
     testTimeout: 30000,
     hookTimeout: 30000,
   },
-});
+})
